Improve reconnection / Control Panel
increased the wait time as we now fetch the iframe url async via api.

Refactor workspace API and update templates

- Added `service` parameter to `view_desktop` endpoint in `workspace.py`.
- Updated iframe source handling based on `service` parameter in `workspace.py`.
- Simplified `view_workspace` route in `pages/workspace.py`.
- Removed async URL loading and share URLs script from `iframe.html`.
- Created new `workspace.html` template for dynamic iframe loading.

Refactor workspace API and update templates

- Added `service` parameter to `view_desktop` endpoint in `workspace.py`.
- Updated iframe source handling based on `service` parameter in `workspace.py`.
- Simplified `view_workspace` route in `pages/workspace.py`.
- Removed async URL loading and share URLs script from `iframe.html`.
- Created new `workspace.html` template for dynamic iframe loading.

Improve reconnection on challenge (re)start

- Added API endpoint to retrieve the current iframe URL.
- Moved `container_password` to `utils`.
- Relocated `start_on_demand_service` to `utils/workspace`.
- Removed special route for workspace desktop.
- Added edge-case handling to the workspace `<service>` route.
- Updated `iframe.html` to support new functionality.
- Modified `navbar.js` to use the new API endpoint.
