+++ dojo_plugin/api/__init__.py
@@ -10,6 +10,7 @@ from .v1.score import score_namespace
 from .v1.scoreboard import scoreboard_namespace
 from .v1.ssh_key import ssh_key_namespace
 from .v1.workspace_tokens import workspace_tokens_namespace
+from .v1.workspace import workspace_namespace
 
 
 api = Blueprint("pwncollege_api", __name__)
@@ -24,3 +25,4 @@ api_v1.add_namespace(score_namespace, "/score")
 api_v1.add_namespace(scoreboard_namespace, "/scoreboard")
 api_v1.add_namespace(ssh_key_namespace, "/ssh_key")
 api_v1.add_namespace(workspace_tokens_namespace, "/workspace_tokens")
+api_v1.add_namespace(workspace_namespace, "/workspace")
