+++ dojo_theme/templates/workspace.html
@@ -1,0 +1,63 @@
+{% extends "base.html" %}
+
+{% block stylesheets %}
+{{ super() }}
+<style>
+    html, body {
+        margin: 0;
+        height: 100%;
+        min-height: 100%;
+    }
+    body {
+        display: flex;
+        flex-direction: column;
+    }
+    nav, footer {
+        flex: none;
+    }
+    main {
+        flex: auto;
+        margin-bottom: 0px;
+    }
+    iframe {
+        height: 100%;
+        width: 100%;
+        overflow: hidden;
+        border: none;
+    }
+    .inactive {
+        height: 50%;
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        text-align: center;
+    }
+</style>
+{% endblock %}
+
+{% block content %}
+
+<iframe id="workspace_iframe" name="{{ iframe_name }}" src="" class="d-none"></iframe>
+<h1 class="inactive">Loading...</h1>
+
+{% endblock %}
+
+{% block scripts %}
+{{ super() }}
+
+<script>
+    let response =  fetch("/pwncollege_api/v1/workspace?service={{ service }}").then(response => response.json()).then(data => {
+        console.log(data);
+        if (data.active) {
+            document.getElementById("workspace_iframe").src = data.iframe_src;
+            document.querySelector("h1").classList.add("d-none");
+            document.getElementById("workspace_iframe").classList.remove("d-none");
+        }
+        else {
+            document.querySelector("h1").innerText = "No active challenge session; start a challenge!";
+        }
+    }).catch(error => {
+        document.querySelector("h1").innerText= `Error while retrieving IFrame Url: ${error}`;
+    });
+</script>
+{% endblock %}
